Initial commit of Rshell project.!
By this initial commit, Rshell is a simple shell utility for Linux.

The features it contains so far are:
* A parser which processes the input from the user to the shell and
  output the parsed command ready to be executed.
* A shell that executes the parsed input from the user.

So far Rshell is capable of run general GNU utilities without problem
such as ls, whoami, pwd, touch, and so on.

It has problems when running bigger programs such as Vim, Radare2 and
other big CLI apps.